requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
pyyaml
dvc
numpy
pandas
scikit-learn
mlflow
python-dotenv>=0.5.1
